Add files via upload
I'd like to use this tutorial for a course I'm teaching, but it's at the more introductory level, so I wanted to add in more information about connecting folders via catalog tab and just how to open/view the attribute table. I'd like to add in a few images to show those steps as well.